Reorganized views; refactored About, EasterEgg, HeadToHead views; start refactoring CampaignList
