+++ Shared Widgets/CampaignTitle.swift
