+++ Shared Widgets/EntryView.swift
