+++ Shared/Views/App/FundraiserListItem.swift
