+++ Shared/Views/App/Other/AnimatedAdaptiveImage.swift
