+++ Shared/Views/Legacy/SettingsDivider.swift
