+++ Shared/Views/Legacy/WidgetSettingsView.swift
